# back-end/src/main/java/com/controle/controle_gastos/application/controller/UserController.java
package com.controle.controle_gastos.application.controller;

import com.controle.controle_gastos.domain.entity.User;
import com.controle.controle_gastos.domain.exceptions.DomainException;
import com.controle.controle_gastos.domain.service.UserService;
import com.controle.controle_gastos.domain.to.PageTO;
import com.controle.controle_gastos.domain.to.PaginationTO;
import com.controle.controle_gastos.domain.to.SummaryResponseTO;
import com.controle.controle_gastos.domain.to.UserTO;
import jakarta.validation.Valid;
import java.util.HashMap;
import java.util.Map;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RequestMapping(value = "/v1/user")
@RestController
public class UserController {

  private final UserService userService;

  public UserController(UserService userService) {
    this.userService = userService;
  }

  @GetMapping(value = "", produces = MediaType.APPLICATION_JSON_VALUE)
  public ResponseEntity<PageTO<User>> getAll(
      @RequestParam(name = "page", defaultValue = "0") int page,
      @RequestParam(name = "size", defaultValue = "10") int size
  ) {
    PaginationTO paginationTO = new PaginationTO(page, size);
    Map<String, Object> params = new HashMap<>();
    paginationTO.setParams(params);

    return ResponseEntity.ok(userService.findAll(paginationTO));
  }

  @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
  public ResponseEntity<User> getById(@Valid @PathVariable("id") long id) throws DomainException {
    return ResponseEntity.ok(userService.findById(id));
  }

  @PostMapping(value = "", produces = MediaType.APPLICATION_JSON_VALUE)
  public ResponseEntity<User> create(@Valid @RequestBody UserTO userTO) throws DomainException {
    return new ResponseEntity<>(userService.createUser(userTO), HttpStatus.CREATED);
  }

  @DeleteMapping(value = "/{id}")
  public ResponseEntity<Object> delete(@PathVariable("id") long id) throws DomainException {
    userService.deleteUser(id);
    return ResponseEntity.noContent().build();
  }

  @GetMapping(value = "/summary", produces = MediaType.APPLICATION_JSON_VALUE)
  public ResponseEntity<SummaryResponseTO> getSummary() throws DomainException {
    return ResponseEntity.ok(userService.getSummary());
  }
}